+++ libsnark/gadgetlib1/tests/test_bls12_377_pairing.cpp
@@ -7,7 +7,6 @@
  *****************************************************************************/
 
 #include "libsnark/gadgetlib1/gadgets/fields/fp12_2over3over2_gadgets.hpp"
-#include "libsnark/gadgetlib1/gadgets/pairing/bw6_761_bls12_377/bls12_377_pairing.hpp"
 #include "libsnark/gadgetlib1/gadgets/pairing/bw6_761_bls12_377/bw6_761_pairing_params.hpp"
 #include "libsnark/gadgetlib1/gadgets/pairing/pairing_params.hpp"
 #include "libsnark/zk_proof_systems/ppzksnark/r1cs_gg_ppzksnark/r1cs_gg_ppzksnark.hpp"
